+++ libs/insight-viewer/src/hooks/useImage/loadImage.spec.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-shadow */
-import { CONFIG, IMAGE_LOADER_SCHEME } from '../../const'
+import { CONFIG } from '../../const'
 import { CornerstoneImage } from '../../utils/cornerstoneHelper'
 import { CORNERSTONE_IMAGE_MOCK } from '../../mocks/const'
 import { loadImage } from './loadImage'
@@ -8,10 +8,8 @@ import { loadCornerstoneImage } from './loadCornerstoneImage'
 const IMAGE_ID = 'wadouri:https://example/CT000000.dcm'
 const defaultParam = {
   onError: CONFIG.onError,
-  requestInterceptor: CONFIG.requestInterceptor,
   imageId: IMAGE_ID,
-  timeout: CONFIG.timeout,
-  imageScheme: IMAGE_LOADER_SCHEME.WADO,
+  loader: undefined,
 }
 const cornerstoneImage = CORNERSTONE_IMAGE_MOCK as unknown as CornerstoneImage
 const mockLoadCornerstoneImage = loadCornerstoneImage as jest.Mock