+++ libs/insight-viewer/src/hooks/useImage/types.ts
@@ -1,13 +1,8 @@
-import { HTTP, RequestInterceptor, ImageLoaderScheme } from '../../types'
+import { HTTP, Loader } from '../../types'
 import { ImageWithoutKey } from '../../Viewer/types'
 
 export type Props = {
   imageId: string
 } & Partial<HTTP>
 
-export type GetImage = (arg: {
-  imageId: string
-  imageScheme: ImageLoaderScheme
-  requestInterceptor: RequestInterceptor
-  timeout: number
-}) => Promise<ImageWithoutKey>
+export type GetImage = (arg: { imageId: string; loader: Loader | undefined }) => Promise<ImageWithoutKey>