+++ libs/insight-viewer/src/hooks/useMultipleImages/loadCornerstoneImages.ts
@@ -4,25 +4,44 @@ import { IMAGE_LOADER_SCHEME } from '../../const'
 import { getHttpClient } from '../../utils/httpClient'
 import { ImageWithoutKey } from '../../Viewer/types'
 
-interface GetImage {
-  (arg: {
-    imageId: string
-    imageScheme: ImageLoaderScheme
-    requestInterceptor: RequestInterceptor
-    timeout: number
-  }): Promise<ImageWithoutKey>
+interface LoadCornerstoneImagesVer1 {
+  imageScheme: ImageLoaderScheme
+  requestInterceptor: RequestInterceptor
+  timeout: number
+  loader?: never
+}
+
+interface LoadCornerstoneImagesVer2 {
+  imageScheme?: never
+  requestInterceptor?: never
+  timeout?: never
+  loader: (url: string) => Promise<ArrayBuffer>
+}
+
+interface LoadCornerstoneImages {
+  (arg: { imageId: string } & (LoadCornerstoneImagesVer1 | LoadCornerstoneImagesVer2)): Promise<ImageWithoutKey>
 }
 
 /**
  * It calls cornerstone.js loadImage. It is pluggable for unit test.
  */
-export const loadCornerstoneImages: GetImage = ({ imageId, imageScheme, requestInterceptor, timeout }) =>
-  loadImage(imageId, {
-    loader:
-      imageScheme === IMAGE_LOADER_SCHEME.DICOMFILE
-        ? undefined
-        : getHttpClient({
-            requestInterceptor,
-            timeout,
-          }),
-  })
+export const loadCornerstoneImages: LoadCornerstoneImages = ({
+  imageId,
+  loader,
+  imageScheme,
+  requestInterceptor,
+  timeout,
+}) => {
+  const doesNeedLoader = imageScheme !== IMAGE_LOADER_SCHEME.DICOMFILE
+
+  if (!doesNeedLoader) {
+    return loadImage(imageId, { loader: undefined })
+  }
+
+  if (loader === undefined) {
+    const httpClient = getHttpClient({ requestInterceptor, timeout })
+    return loadImage(imageId, { loader: httpClient })
+  }
+
+  return loadImage(imageId, { loader })
+}