+++ libs/insight-viewer/src/utils/common/getMeasurement.ts
@@ -1,17 +1,19 @@
 import { getLineLength } from './getLineLength'
+
+import { getCircleCenterPoint } from './getCircleCenterPoint'
 import { getCircleRadiusByMeasuringUnit, getCircleRadius } from './getCircleRadius'
 
 import type { Image } from '../../Viewer/types'
 import type { Point, Measurement, MeasurementMode } from '../../types'
 
 export function getMeasurement(
-  points: [mouseDownPoint: Point, mouseUpPoint: Point],
+  points: [Point, Point],
   textPoint: Point | null,
   mode: MeasurementMode,
   measurements: Measurement[],
   image: Image | null
 ): Measurement {
-  const [mouseDownPoint, mouseUpPoint] = points
+  const [startPoint, endPoint] = points
   const currentId = measurements.length === 0 ? 1 : Math.max(...measurements.map(({ id }) => id), 0) + 1
 
   const defaultMeasurementInfo: Pick<Measurement, 'id' | 'lineWidth'> = {
@@ -20,13 +22,14 @@ export function getMeasurement(
   }
 
   if (mode === 'circle') {
-    const radiusWithoutUnit = getCircleRadius(mouseDownPoint, mouseUpPoint)
-    const { radius, unit } = getCircleRadiusByMeasuringUnit(mouseDownPoint, mouseUpPoint, image)
+    const centerPoint = getCircleCenterPoint(startPoint, endPoint)
+    const radiusWithoutUnit = getCircleRadius(startPoint, endPoint)
+    const { radius, unit } = getCircleRadiusByMeasuringUnit(startPoint, endPoint, image)
 
     return {
       ...defaultMeasurementInfo,
       type: 'circle',
-      centerPoint: mouseDownPoint,
+      centerPoint: centerPoint,
       textPoint,
       radius: radiusWithoutUnit,
       measuredValue: radius,
@@ -35,12 +38,12 @@ export function getMeasurement(
   }
 
   // Ruler EditMode
-  const { length, unit } = getLineLength(mouseDownPoint, mouseUpPoint, image)
+  const { length, unit } = getLineLength(startPoint, endPoint, image)
   return {
     ...defaultMeasurementInfo,
     type: 'ruler',
     textPoint,
-    startAndEndPoint: [mouseDownPoint, mouseUpPoint],
+    startAndEndPoint: [startPoint, endPoint],
     measuredValue: length,
     unit,
   }