+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [5.9.0](https://github.com/lunit-io/frontend-components/compare/@lunit/insight-viewer@5.8.3...@lunit/insight-viewer@5.9.0) (2022-09-21)
+
+### Features
+
+- add circle center calculation ([aa47736](https://github.com/lunit-io/frontend-components/commit/aa477360946fe7017b323f26001fdcc23465e6ca))
+- add getCircleRadiusByCenter logic ([5c8d6b3](https://github.com/lunit-io/frontend-components/commit/5c8d6b3b81bac86aae68d082742935eabf3af7f1))
+- change centerPoint generate logic ([90adf0f](https://github.com/lunit-io/frontend-components/commit/90adf0fdc851ba24faea86cbda380484d8f7ade0))
+- change drawingRadius function of circle ([1c269d3](https://github.com/lunit-io/frontend-components/commit/1c269d32fcc6d9c029fe27ae605cfef81fbd3a30))
+- change edit mode handling logic of circle ([4db1ca4](https://github.com/lunit-io/frontend-components/commit/4db1ca4d014eeef1c256d8670d98fef2b14c6f56))
+- change points handling logic of circle ([d0107c8](https://github.com/lunit-io/frontend-components/commit/d0107c8f2f19af61c4a0d48468896edd279c01e2))
+- **docs:** add image selector on annotation and measurement ([b2cb1f0](https://github.com/lunit-io/frontend-components/commit/b2cb1f0295ca1811e403c49b7d4a56b5bde80a81))
+- **docs:** add usage for new useImage in Basic page ([ac2fb6e](https://github.com/lunit-io/frontend-components/commit/ac2fb6ed970175586a7a390f44d9dc2fa3e0783d))
+- **viewer:** add getCircleStartPoint function ([8bbf73b](https://github.com/lunit-io/frontend-components/commit/8bbf73bb4a3e276d969b1e760559cb2122c92e9a))
+- **viewer:** add validation util functions of points ([8477751](https://github.com/lunit-io/frontend-components/commit/847775162adf6b7b585b6d2d9850b00e7cd4e4f1))
+- **viewer:** apply validation of same points when drawing ([3d9592f](https://github.com/lunit-io/frontend-components/commit/3d9592febb1b767b3d929d92dbbb6382fe03048d))
+- **viewer:** change ruler text position ([17f6387](https://github.com/lunit-io/frontend-components/commit/17f6387e48fb8301e7cbc7574e34468bea3830ab))
+- **viewer:** modify getRulerTextPosition ([62ca329](https://github.com/lunit-io/frontend-components/commit/62ca329e4ca788908f577aa59265c2028b30170b))
+- **viewer:** modify validation logic to Annotation, Measurement ([51e2c79](https://github.com/lunit-io/frontend-components/commit/51e2c79777b67736a07ca69b6b2f463559bfd43d))
+- **viewer:** modify zoom feature ([71091e0](https://github.com/lunit-io/frontend-components/commit/71091e0cde2cb8a3a8e63efccb634995d531b7f7))
+- **viewer:** parameterize useImage's HTTP client ([65fc366](https://github.com/lunit-io/frontend-components/commit/65fc36624a1aa93e64b34cbe4b314e269f941ab2))
+
+### Bug Fixes
+
+- **docs:** remove not exist type export ([ee4667a](https://github.com/lunit-io/frontend-components/commit/ee4667a9418437c3db90ae1978be1955f070cf73))
+- **viewer:** add drawingMode to getEditPointPosition ([5b51d8c](https://github.com/lunit-io/frontend-components/commit/5b51d8c9ffd1182190ad9a6d372455ca8e1a8aa3))
+- **viewer:** add isStartedFromCenter to calculateRadius ([8189d04](https://github.com/lunit-io/frontend-components/commit/8189d04a53f4caf42de8f62377a4a80be0e73b84))
+- **viewer:** change wrong constant name ([760be30](https://github.com/lunit-io/frontend-components/commit/760be306b997e76cd6770a219b75d38281eba1e1))
+- **viewer:** modify conditional statement ([4eb9e07](https://github.com/lunit-io/frontend-components/commit/4eb9e0740f903a21923fd97243bc908711ac43e8))
+- **viewer:** scale logic ([760857d](https://github.com/lunit-io/frontend-components/commit/760857dfb85e9205871fa3ea6f3475c27d64b535))
+- **viewer:** text default position ([3dd8ecb](https://github.com/lunit-io/frontend-components/commit/3dd8ecbcbc6f0dac2389416f665b6e9aed85a62f))
+
 ## [5.8.3](https://github.com/lunit-io/frontend-components/compare/@lunit/insight-viewer@5.8.2...@lunit/insight-viewer@5.8.3) (2022-09-16)
 
 ### Features